Added -p/--paths option to set list of paths
The paths argument is a file name that contains a list of paths to
request to the server. The any non-empty line that doesn't start with
'#' must be a path-only URL. The main URL still needs to be set in oder
to gather the host and port information. When using a local socket,
however, the main URL may be omitted.