script: add ability to exit thread after script call
This allows the request and response hooks to cause the thread to cease
sending new requests, useful when we would like to end our run without
respect to time.

The request function exits by returning a nil value; the response function
signals an exit with a boolean true/false return.