add support for multilib system dependencies
On some systems where both 32bit and 64bit versions of a library can
coexist, it is incorrect to assume lib/ for unbundled dependencies.

The $(LIB) variable defaults to "lib" and can for example be changed
to "lib64" on Enterprise Linux and derivatives. The name LIB matches
RPM's convention and could be used in conjunction with the %{_lib}
macro:

    %{make_build} WITH_OPENSSL=/usr WITH_LUAJIT=/usr LIB=%{_lib}

This should ease ports to other platforms.