make the wrk.headers table case-insensitive
In the HTTP grammar header fields names are case-insensitive, so adding
-Hhost:example.com on the command line would result in a request with
two host headers, leading to a 400 error on web servers that enforce
that HTTP/1.1 requests can only have one host header:

    GET / HTTP/1.1
    Host: 127.0.0.1:8080
    host: example.com

To work around the fact that keys in a lua table are case-sensitive,
the wrk.headers field is wrapped with a case-insensitive metatable. If
a user script passes a headers table, it is also deduplicated via a temp
table since iterating over pairs may result in multiple cases for header
names.

It is possible to have multiple header fields with the same name in an
HTTP request but the recommendation is generally to combine them. It
doesn't work with all headers (for example Accept works, Host doesn't)
and considering how Lua's table works, it's probably simpler to give
up the ability to have multiple headers with the same name to avoid
unintended duplicates.